trading-bot/requirements.txt
ccxt==4.3.82
streamlit==1.32.0
pydantic==1.10.14
python-dotenv==1.0.1
web3==6.11.1
requests==2.31.0
aiohttp==3.9.3
tenacity==8.2.3
